Define NotionBlocks
