+++ src/components/blog-parts.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import Link from 'next/link'
 
-import NotionBlock from './notion-block'
-import * as interfaces from '../lib/notion/interfaces'
+import NotionBlocks from './notion-block'
 import {
   getBeforeLink,
   getBlogLink,
@@ -54,9 +53,7 @@ export const PostExcerpt = ({ post }) => (
 
 export const PostBody = ({ blocks }) => (
   <div className={styles.postBody}>
-    {wrapListItems(blocks).map((block, i) => (
-      <NotionBlock block={block} key={`post-body-${i}`} />
-    ))}
+    <NotionBlocks blocks={blocks} />
   </div>
 )
 
@@ -156,38 +153,3 @@ export const PostsNotFound = () => (
     Woops! did not find the posts, redirecting you back to the blog index
   </div>
 )
-
-const wrapListItems = blocks =>
-  blocks.reduce((arr, block, i) => {
-    const isBulletedListItem = block.Type === 'bulleted_list_item'
-    const isNumberedListItem = block.Type === 'numbered_list_item'
-
-    if (!isBulletedListItem && !isNumberedListItem) return arr.concat(block)
-
-    const listType = isBulletedListItem ? 'bulleted_list' : 'numbered_list'
-
-    if (i === 0) {
-      const list: interfaces.List = {
-        Type: listType,
-        ListItems: [block],
-      }
-      return arr.concat(list)
-    }
-
-    const prevList = arr[arr.length - 1]
-
-    if (
-      (isBulletedListItem && prevList.Type !== 'bulleted_list') ||
-      (isNumberedListItem && prevList.Type !== 'numbered_list')
-    ) {
-      const list: interfaces.List = {
-        Type: listType,
-        ListItems: [block],
-      }
-      return arr.concat(list)
-    }
-
-    prevList.ListItems.push(block)
-
-    return arr
-  }, [])
